Index.html updates
changes file names and variables for sign in and sign out to log in and
log out curl scripts.
Added event handlers for change password and updated login and out
handlers in app.js
changed variable names from signin and out to login and out in
events.js.
updated the success functions from sign in and out to log in and out,
also added the changePasswordSuccess and failure functions.